fix data type number 0f floor in feature

+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/UpdateLandAndBuildingProperty/UpdateLandAndBuildingPropertyCommand.cs
@@ -123,7 +123,7 @@ public record UpdateLandAndBuildingPropertyCommand(
     // Building Info
     string? BuildingType = null,
     string? BuildingTypeOther = null,
-    int? NumberOfFloors = null,
+    decimal? NumberOfFloors = null,
     string? DecorationType = null,
     string? DecorationTypeOther = null,
     bool? IsEncroachingOthers = null,
